pyproject.toml
[tool.poetry]
name = "myOCR"
version = "0.1.0"
description = ""
authors = ["genecoding"]

[tool.poetry.dependencies]
python = "~3.7"
click = "^8.1.0"
opencv-python = "~4.1.2"
pytesseract = "^0.3.9"
pdf2image = "^1.16.0"
pyenchant = "^3.2.2"
nltk = "^3.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
